pyproject.toml
[tool.poetry]
name = "videobot"
version = "1.1.0"
description = "Bot that makes video creation so simple"
authors = ["Drugsosos <44712637+Drugsosos@users.noreply.github.com>"]
license = "Apache-2.0"

[tool.poetry.dependencies]
python = "^3.10"
wheel = "0.37.1"
python-dotenv = "0.20.0"
asyncpraw = "7.5.0"
aiohttp = "3.8.1"
pyppeteer = "1.0.2"
MoviePy = "1.0.3"
tqdm = "4.64.0"
rich = "12.4.4"
aiofiles = "0.6.0"
pytube = "12.1.0"
youtube-search-python = "1.6.5"
requests = "^2.28.0"
mutagen = "^1.45.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
